refactoring: правки после ревю вынес константы в application.yaml обработка ошибок

+++ src/main/java/faang/school/paymentservice/dto/ExchangeRatesResponse.java
@@ -1,8 +1,9 @@
 package faang.school.paymentservice.dto;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 public record ExchangeRatesResponse(
         String base,
-        Map<String, Double> rates
+        Map<String, BigDecimal> rates
 ) {}
